Migrated Tangent Mode to separate files
