+++ Math.py
@@ -1,12 +1,6 @@
-import findspark
-findspark.init()
-from pyspark import SparkConf, SparkContext
-
 import numpy as np
 from collections import Sequence
 
-conf = SparkConf().setMaster("local").setAppName("Differentiation")
-sc = SparkContext.getOrCreate(conf = conf)
 import math
 
 
@@ -88,74 +82,3 @@ def pow(x, y):
     if not isinstance(y, Tangent_Type):
             y = Tangent_Type(y)
     return Tangent_Type(math.pow(x.v, y.v), x.t*y.v*x.v**(y.v-1) + math.log(x.v)*y.t*x.v**y.v)
-
-
-class Tangent_Mode:
-    def __init__(self, n_input, n_output, f):
-        self.n_input = n_input
-        self.n_output = n_output
-        self.f = f
-    
-    def computeFullJacobian(self, x):
-        x = list(map(Tangent_Type, x))
-        Jacobian = np.zeros(shape=(self.n_output, self.n_input), dtype="float64")
-        for i,xi in enumerate(x):
-            xi.t = 1.0
-            y = self.f(x)
-            if isinstance(y, Sequence):
-                for j in range(self.n_output):
-                    Jacobian[j,i] = y[j].t
-            else:
-                Jacobian[0,i] = y.t
-            xi.t = 0.0
-        return Jacobian, y.v
-    
-    def computeDerivative(self, x, x_t):
-        x = list(map(Tangent_Type, x))
-        for i, xi_t in enumerate(x_t):
-            x[i].t = xi_t
-        directionalDerivative = np.zeros(shape=(self.n_output,), dtype="float64")
-        y = self.f(x)
-        if isinstance(y, Sequence):
-            for j in range(self.n_output):
-                directionalDerivative[j] = y[j].t
-        else:
-            directionalDerivative[0] = y.t
-        return directionalDerivative, y.v
-
-
-class Tangent_Mode_Distributed:
-    def __init__(self, n_input, n_output, f):
-        self.n_input = n_input
-        self.n_output = n_output
-        self.f = f
-    
-    def computeFullJacobian(self, x):
-        self.x = list(map(Tangent_Type, x))
-        Jacobian = np.zeros(shape=(self.n_output, self.n_input), dtype="float64")
-        x_t = [0]*self.n_input
-        tangent_rdd = sc.parallelize([])
-        for i in range(len(x)):
-            x_t[i] = 1
-            tangent_rdd=tangent_rdd.union(sc.parallelize([x_t]))
-            x_t[i] = 0
-        
-        jacobian_rdd = tangent_rdd.map(self.computeDerivative)
-        Jacobian = jacobian_rdd.collect()
-        return Jacobian
-    
-    def computeDerivative(self, x_t):
-        loc_one = -1
-        for i, xi_t in enumerate(x_t):
-            if xi_t == 1:
-                loc_one = i + 1
-            self.x[i].t = xi_t
-        directionalDerivative = np.zeros(shape=(self.n_output,), dtype="float64")
-        y = self.f(self.x)
-        if isinstance(y, Sequence):
-            for j in range(self.n_output):
-                directionalDerivative[j] = y[j].t
-        else:
-            directionalDerivative[0] = y.t
-        return loc_one, directionalDerivative
-
